combine task pages controllers
